fix(acessos): id_entrada é bigint, não integer
clientes_entrada_new.id_entrada é int8 (bigint) no banco. A RPC
declarava `id_entrada integer` no RETURNS TABLE, o que disparava
"structure of query does not match function result type" em runtime.

+++ supabase-migrations/20260421_create_client_access_overview_rpc.sql
@@ -9,7 +9,7 @@ DROP FUNCTION IF EXISTS public.get_client_access_overview();
 
 CREATE OR REPLACE FUNCTION public.get_client_access_overview()
 RETURNS TABLE (
-  id_entrada integer,
+  id_entrada bigint,
   id_cliente uuid,
   nome_cliente text,
   nome_empresa text,
